requirements.txt
absl-py==2.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dm-tree==0.1.8
docstring_parser==0.16
etils==1.11.0
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
immutabledict==4.2.1
importlib_resources==6.4.5
keras==3.7.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
minio==7.2.14
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
promise==2.3
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
simple-parsing==0.1.6
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
